refactor: Move LocationFetcher to view_location_fetcher.dart
