+++ lib/screens/LocationsOverviewScreen.dart
@@ -25,6 +25,7 @@ import 'package:locus/screens/SharesOverviewScreen.dart';
 import 'package:locus/screens/locations_overview_screen_widgets/ActiveSharesSheet.dart';
 import 'package:locus/screens/locations_overview_screen_widgets/OutOfBoundMarker.dart';
 import 'package:locus/screens/locations_overview_screen_widgets/ShareLocationSheet.dart';
+import 'package:locus/screens/locations_overview_screen_widgets/view_location_fetcher.dart';
 import 'package:locus/services/task_service.dart';
 import 'package:locus/services/view_service.dart';
 import 'package:locus/utils/helpers.dart';
@@ -66,100 +67,6 @@ enum LocationStatus {
   fetching,
 }
 
-class LocationFetcher extends ChangeNotifier {
-  final Iterable<TaskView> views;
-  final Map<TaskView, List<LocationPointService>> _locations = {};
-  final List<VoidCallback> _getLocationsUnsubscribers = [];
-
-  bool _mounted = true;
-
-  Map<TaskView, List<LocationPointService>> get locations => _locations;
-
-  bool _isLoading = false;
-
-  bool get isLoading => _isLoading;
-
-  LocationFetcher(this.views);
-
-  bool get hasMultipleLocationViews => _locations.keys.length > 1;
-
-  // If _fetchLast24Hours fails (no location fetched), we want to get the last location
-  void _fetchLastLocation(final TaskView view) {
-    _getLocationsUnsubscribers.add(
-      view.getLocations(
-        limit: 1,
-        onLocationFetched: (location) {
-          if (!_mounted) {
-            return;
-          }
-
-          _locations[view] = [location];
-        },
-        onEnd: () {
-          if (!_mounted) {
-            return;
-          }
-
-          _setIsLoading(_locations.keys.length == views.length);
-        },
-      ),
-    );
-  }
-
-  void _fetchLast24Hours() {
-    _getLocationsUnsubscribers.addAll(
-      views.map(
-        (view) => view.getLocations(
-          from: DateTime.now().subtract(const Duration(days: 1)),
-          onLocationFetched: (location) {
-            if (!_mounted) {
-              return;
-            }
-
-            _locations[view] = List<LocationPointService>.from(
-              [..._locations[view] ?? [], location],
-            );
-          },
-          onEnd: () {
-            if (!_mounted) {
-              return;
-            }
-
-            _locations[view] = _locations[view]!
-              ..sort((a, b) => a.createdAt.compareTo(b.createdAt));
-
-            _setIsLoading(_locations.keys.length == views.length);
-          },
-          onEmptyEnd: () {
-            _fetchLastLocation(view);
-          },
-        ),
-      ),
-    );
-  }
-
-  void fetchLocations() {
-    _setIsLoading(true);
-
-    _fetchLast24Hours();
-  }
-
-  void _setIsLoading(final bool isLoading) {
-    _isLoading = isLoading;
-    notifyListeners();
-  }
-
-  @override
-  void dispose() {
-    for (final unsubscribe in _getLocationsUnsubscribers) {
-      unsubscribe();
-    }
-
-    _mounted = false;
-    super.dispose();
-  }
-}
-
 class LocationsOverviewScreen extends StatefulWidget {
   const LocationsOverviewScreen({super.key});
 
@@ -173,7 +80,7 @@ class _LocationsOverviewScreenState extends State<LocationsOverviewScreen>
         AutomaticKeepAliveClientMixin,
         WidgetsBindingObserver,
         TickerProviderStateMixin {
-  late final LocationFetcher _fetchers;
+  late final ViewLocationFetcher _fetchers;
   MapController? flutterMapController;
   AppleMaps.AppleMapController? appleMapController;
 
@@ -328,7 +235,7 @@ class _LocationsOverviewScreenState extends State<LocationsOverviewScreen>
   void _createLocationFetcher() {
     final viewService = context.read<ViewService>();
 
-    _fetchers = LocationFetcher(viewService.views)..fetchLocations();
+    _fetchers = ViewLocationFetcher(viewService.views)..fetchLocations();
   }
 
   void _rebuild() {
